src/pali/pyproject.toml
[project]
name = "pali"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.1",
    "fastapi>=0.125.0",
    "google-cloud-aiplatform>=1.132.0",
    "pandas>=2.3.3",
    "rapidfuzz>=3.14.3",
    "uvicorn[standard]>=0.38.0",
]
